+++ src/main/java/si/matjazcerkvenik/alertmonitor/util/AmMetrics.java
@@ -10,7 +10,7 @@ public class AmMetrics {
 
     public static final Gauge alertmonitor_build_info = Gauge.build()
             .name("alertmonitor_build_info")
-            .help("Alertmonitor version, value is start time")
+            .help("Alertmonitor version, value contains start time")
             .labelNames("app", "version", "os")
             .register();
 
@@ -46,6 +46,11 @@ public class AmMetrics {
             .name("alertmonitor_resync_task_total")
             .help("Total number of executed resync tasks.")
             .labelNames("result")
+            .register();
+
+    public static final Gauge alertmonitor_resync_interval_seconds = Gauge.build()
+            .name("alertmonitor_resync_interval_seconds")
+            .help("Resync interval.")
             .register();
 
 }